Tidy up tests.
