+++ test/Figure.test.js
@@ -1,0 +1,19 @@
+import { test } from 'substance-test'
+import { setCursor, openManuscriptEditor, PseudoFileEvent } from './shared/integrationTestHelpers'
+import setupTestApp from './shared/setupTestApp'
+
+test('Figure: add figure into manuscript', t => {
+  let { app } = setupTestApp(t, { archiveId: 'blank' })
+  let editor = openManuscriptEditor(app)
+  setCursor(editor, 'p-2.content', 5)
+  // ATTENTION: it is not possible to trigger the file-dialog programmatically
+  // instead we are just checking that this does not throw
+  let insertFigureTool = editor.find('.sc-insert-figure-tool')
+  t.ok(insertFigureTool.find('button').click(), 'clicking on the insert figure button should not throw error')
+  // ... and then triggering onFileSelect() directly
+  insertFigureTool.onFileSelect(new PseudoFileEvent())
+  let afterP2 = editor.find('*[data-id=p-2] + *')
+  t.ok(afterP2.hasClass('sm-figure'), 'element after p-2 should be a figure now')
+  // TODO: we should test the automatic labeling
+  t.end()
+})
