+++ test/Table.test.js
@@ -5,9 +5,18 @@ import {
 } from '../index'
 import { getMountPoint, DOMEvent } from './shared/testHelpers'
 import setupTestArticleSession from './shared/setupTestArticleSession'
+import { openManuscriptEditor, loadBodyFixture, getDocument, setSelection, getApi, getEditorSession } from './shared/integrationTestHelpers'
+import setupTestApp from './shared/setupTestApp'
+
+// TODO: add tests for table cells
+// - add ext-link
+// - cite figure
+// - cite table
+// - cite table-footnotes
+// - cite reference
 
 test('Table: mounting a table component', t => {
-  let { table, context } = _setup(t)
+  let { table, context } = _setupEditorWithOneTable(t)
   let el = getMountPoint(t)
   let comp = new TableComponent(null, { node: table }, { context })
   comp.mount(el)
@@ -16,7 +25,7 @@ test('Table: mounting a table component', t => {
 })
 
 test('Table: setting table selections', t => {
-  let { editorSession, table, context } = _setup(t)
+  let { editorSession, table, context } = _setupEditorWithOneTable(t)
   let el = getMountPoint(t)
   let comp = new TableComponent(null, { node: table }, { context })
   comp.mount(el)
@@ -51,7 +60,7 @@ test('Table: setting table selections', t => {
 })
 
 test('Table: mouse interactions', t => {
-  let { editorSession, table, context } = _setup(t)
+  let { editorSession, table, context } = _setupEditorWithOneTable(t)
   let el = getMountPoint(t)
   let comp = new TableComponent(null, { node: table }, { context })
   comp.mount(el)
@@ -83,7 +92,7 @@ test('Table: mouse interactions', t => {
 })
 
 test('Table: keyboard interactions', t => {
-  let { editorSession, table, context } = _setup(t)
+  let { editorSession, table, context } = _setupEditorWithOneTable(t)
   let el = getMountPoint(t)
   let comp = new TableComponent(null, { node: table }, { context })
   comp.mount(el)
@@ -128,7 +137,161 @@ test('Table: keyboard interactions', t => {
   t.end()
 })
 
-function _setup (t) {
+const SIMPLE_TABLE = `<table-wrap>
+  <table>
+    <tbody>
+      <tr>
+        <td id="t11">aaa</td>
+        <td id="t12">bbb</td>
+        <td id="t13">ccc</td>
+      </tr>
+      <tr>
+        <td id="t21">ddd</td>
+        <td id="t22">eee</td>
+        <td id="t23">fff</td>
+      </tr>
+      <tr>
+        <td id="t31">ggg</td>
+        <td id="t32">hhh</td>
+        <td id="t33">iii</td>
+      </tr>
+    </tbody>
+  </table>
+</table-wrap>`
+
+test('Table: formatting in table cells', t => {
+  let { app } = setupTestApp(t, { archiveId: 'blank' })
+  let editor = openManuscriptEditor(app)
+  let doc = getDocument(editor)
+  loadBodyFixture(editor, SIMPLE_TABLE)
+
+  // use bold tool
+  setSelection(editor, 't11.content', 1, 2)
+  let boldTool = editor.find('.sc-toggle-tool.sm-bold')
+  // click on list tool to turn "p1" into a list
+  boldTool.find('button').el.click()
+  let annos = doc.getAnnotations(['t11', 'content'])
+  t.deepEqual(annos.map(a => a.type), ['bold'], 'there should be one bold annotation on t11')
+
+  // use italic tool
+  setSelection(editor, 't12.content', 1, 2)
+  let italicTool = editor.find('.sc-toggle-tool.sm-italic')
+  // click on list tool to turn "p1" into a list
+  italicTool.find('button').el.click()
+  annos = doc.getAnnotations(['t12', 'content'])
+  t.deepEqual(annos.map(a => a.type), ['italic'], 'there should be one italic annotation on t12')
+
+  // use sup tool
+  setSelection(editor, 't13.content', 1, 2)
+  let supTool = editor.find('.sc-toggle-tool.sm-sup')
+  // click on list tool to turn "p1" into a list
+  supTool.find('button').el.click()
+  annos = doc.getAnnotations(['t13', 'content'])
+  t.deepEqual(annos.map(a => a.type), ['sup'], 'there should be one sup annotation on t13')
+
+  // use sub tool
+  setSelection(editor, 't21.content', 1, 2)
+  let subTool = editor.find('.sc-toggle-tool.sm-sub')
+  // click on list tool to turn "p1" into a list
+  subTool.find('button').el.click()
+  annos = doc.getAnnotations(['t21', 'content'])
+  t.deepEqual(annos.map(a => a.type), ['sub'], 'there should be one sub annotation on t21')
+
+  // use monospace tool
+  setSelection(editor, 't22.content', 1, 2)
+  let monospaceTool = editor.find('.sc-toggle-tool.sm-monospace')
+  // click on list tool to turn "p1" into a list
+  monospaceTool.find('button').el.click()
+  annos = doc.getAnnotations(['t22', 'content'])
+  t.deepEqual(annos.map(a => a.type), ['monospace'], 'there should be one monospace annotation on t22')
+
+  t.end()
+})
+
+const TABLES_AND_REFS = `<p />
+<p id="p1">Foo Bar <xref ref-type="table" rid="t1"></xref></p>
+<table-wrap id="t1">
+<table />
+</table-wrap>
+<p id="p2">Lorem ipsum <xref ref-type="table" rid="t2"></xref></p>
+<table-wrap id="t2">
+<table />
+</table-wrap>
+<p>Bla bla</p>`
+
+// testing the general ability to insert tables but also look into table citations
+test('Table: inserting and deleting a table into manuscript', t => {
+  let { app } = setupTestApp(t, { archiveId: 'blank' })
+  let editor = openManuscriptEditor(app)
+  let api = getApi(editor)
+  let doc = getDocument(editor)
+  loadBodyFixture(editor, TABLES_AND_REFS)
+
+  let t1 = doc.get('t1')
+  t.equal(t1.state.label, 'Table 1', 't1 should have correct label')
+  let t2 = doc.get('t2')
+  t.equal(t2.state.label, 'Table 2', 't2 should have correct label')
+
+  let p1 = doc.get('p1')
+  let t1ref = p1.find('xref')
+  t.equal(t1ref.state.label, 'Table 1', 'citation of t1 should have correct label')
+  let p2 = doc.get('p2')
+  let t2ref = p2.find('xref')
+  t.equal(t2ref.state.label, 'Table 2', 'citation of t2 should have correct label')
+
+  api._setSelection({
+    type: 'node',
+    nodeId: 't1',
+    containerId: 'body'
+  })
+  api.deleteSelection()
+
+  t.equal(t2.state.label, 'Table 1', 'citation of t2 should have correct label')
+  t1ref = p1.find('xref')
+  t.equal(t1ref.state.label, '???', 'citation of t1 should have correct label')
+  t2ref = p2.find('xref')
+  t.equal(t2ref.state.label, 'Table 1', 'citation of t2 should have correct label')
+
+  // now assing the broken xref to t2
+  let t1refComp = editor.find(`[data-id="${t1ref.id}"]`)
+  t1refComp.el.click()
+
+  let option = editor.find('.sc-edit-xref-tool .se-option')
+  option.el.click()
+
+  t1ref = p1.find('xref')
+  t.equal(t1ref.state.label, 'Table 1', 'citation should now point to t2')
+
+  t.end()
+})
+
+test('Table: selecting a table', t => {
+  let { app } = setupTestApp(t, { archiveId: 'blank' })
+  let editor = openManuscriptEditor(app)
+  let editorSession = getEditorSession(editor)
+  editorSession.transaction(tx => {
+    let body = tx.get('body')
+    let table = tableHelpers.generateTable(tx, 10, 5, 'test-table')
+    body.append(table)
+  })
+  // a click (somewhere) on the isolated node should select the node
+  let isolatedNode = editor.find('[data-id="test-table"]')
+  isolatedNode.el.click()
+  let sel = editorSession.getSelection()
+  t.equal(sel.type, 'node', 'the selection should be a node selection')
+  t.equal(sel.nodeId, 'test-table', '.. pointing to the inserted table should be selected')
+  // but a click inside the table, e.g. on a cell should select the table cell
+  let table = isolatedNode.find('.sc-table')
+  let td = isolatedNode.find('td')
+  table._onMousedown(new DOMEvent({ target: td.el }))
+
+  td.el.click()
+  sel = editorSession.getSelection()
+  t.equal(sel.customType, 'table', 'the selection should be a table selection')
+  t.end()
+})
+
+function _setupEditorWithOneTable (t) {
   let table
   let res = setupTestArticleSession({
     seed: doc => {
