+++ test/index.js
@@ -7,8 +7,8 @@ import './Annotations.test'
 import './AddReference.test'
 import './BodyConverter.test'
 import './Card.test'
-import './Citations.test'
 import './ClipboardNew.test'
+import './CrossReference.test'
 import './EditReference.test'
 import './Footnote.test'
 import './FormulaConverter.test'
